+++ packages/eslint-plugin/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- Added a `require-media-dimensions` lint rule to `@shopify/eslint-plugin` [[#323](https://github.com/Shopify/web-configs/pull/323)]
 
 ## 41.1.0 - 2022-02-01